Move validation functions into func.php and use the same validation in all cases.
